+++ lab2/src/main/java/com/assignment/lab2/dao/EmployerDao.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.assignment.lab2.entity.*;
 
 @Repository
-public interface EmployerDao extends JpaRepository<EmployerEntity,Integer> {
+public interface EmployerDao extends JpaRepository<EmployerEntity,Long> {
 
 }
